Refs #23322 Python 3 compatibility.
