+++ scripts/test/Muon/run_string_utils_test.py
@@ -39,7 +39,7 @@ class RunStringUtilsListToStringTest(unittest.TestCase):
         run_list = [i for i in range(150)]
         with self.assertRaises(IndexError) as context:
             utils.run_list_to_string(run_list)
-        self.assertTrue("Too many runs (150) must be <100" in context.exception)
+            self.assertTrue("Too many runs (150) must be <100" in context.exception)
 
 
 class RunStringUtilsStringToListTest(unittest.TestCase):
@@ -89,14 +89,14 @@ class RunStringUtilsStringToListTest(unittest.TestCase):
         for run_string in run_strings:
             with self.assertRaises(IndexError) as context:
                 utils.run_string_to_list(run_string)
-            self.assertTrue(run_string + " is not a valid run string" in context.exception)
+                self.assertTrue(run_string + " is not a valid run string" in context.exception)
 
     def test_run_string_to_list_throws_for_incorrectly_placed_delimiter(self):
         run_strings = [",1,2,3", "1,2,3,"]
         for run_string in run_strings:
             with self.assertRaises(IndexError) as context:
                 utils.run_string_to_list(run_string)
-            self.assertTrue(run_string + " is not a valid run string" in context.exception)
+                self.assertTrue(run_string + " is not a valid run string" in context.exception)
 
     def test_run_string_to_list_handles_non_consecutive_runs(self):
         run_lists = [[1, 3, 4, 5], [0, 1, 2, 3, 8, 9, 199, 200]]
